requirements.txt
streamlit
pandas
python-dotenv
thefuzz
anthropic
python-Levenshtein
requests
PyMuPDF
PyPDF2